add lint & lint report
